Update the GeoIpOverride Table
- Uses the new metric handles and labels
- Refactors the design to be cleaner
