+++ web_ui/frontend/app/director/components/index.tsx
@@ -1,4 +1,3 @@
 export * from './DirectorCard';
 export * from './DirectorCardList';
 export * from './NamespaceCard';
-export { default as GeoIpErrorTable } from './GeoIpErrorTable';
