+++ web_ui/frontend/components/GeoIpErrorDisplay/StyledTableCell.tsx
@@ -1,0 +1,9 @@
+import { TableCell, styled, tableCellClasses } from '@mui/material';
+
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  [`&.${tableCellClasses.head}`]: {
+    backgroundColor: theme.palette.warning.main,
+  },
+}));
+
+export default StyledTableCell;
