+++ web_ui/frontend/components/GeoIpErrorDisplay/index.ts
@@ -1,0 +1,1 @@
+export {default} from './GeoIpErrorDisplay'
